+++ Quotations Board Backend/DTOs/Bonds/BondTradeAverageStatistic.cs
@@ -3,10 +3,10 @@ public class BondTradeAverageStatistic
     public string BondName { get; set; } = null!;
     public string IssueNo { get; set; } = null!;
     public string BondId { get; set; } = null!;
-    public decimal AverageWeightedTradeYield { get; set; }
-    public decimal TradedVolume { get; set; }
+    public double AverageWeightedTradeYield { get; set; }
+    public double TradedVolume { get; set; }
     public decimal NumberofTrades { get; set; }
-    public decimal WeightedTradeBuyYield { get; set; }
-    public decimal WeightedTradeSellYield { get; set; }
+    public double WeightedTradeBuyYield { get; set; }
+    public double WeightedTradeSellYield { get; set; }
     public double YearsToMaturity { get; set; }
 }
