+++ Quotations Board Backend/DTOs/ImpliedYield/ComputedImpliedYield.cs
@@ -2,10 +2,10 @@ public class ComputedImpliedYield
 {
     public string BondId { get; set; } = null!;
     public DateTime YieldDate { get; set; }
-    public decimal Yield { get; set; }
+    public double Yield { get; set; }
     public string ReasonForSelection { get; set; } = null!;
     public decimal SelectedYield { get; set; }
-    public decimal TradedYield { get; set; }
-    public decimal QuotedYield { get; set; }
-    public decimal PreviousYield { get; set; }
+    public double TradedYield { get; set; }
+    public double QuotedYield { get; set; }
+    public double PreviousYield { get; set; }
 }
