+++ Quotations Board Backend/Models/Quotations/Quotation.cs
@@ -6,10 +6,10 @@ public class Quotation
     [Key]
     [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
     public string Id { get; set; } = null!;
-    public decimal BuyingYield { get; set; }
-    public decimal SellingYield { get; set; }
-    public decimal BuyVolume { get; set; } = 50000000;
-    public decimal SellVolume { get; set; } = 50000000;
+    public double BuyingYield { get; set; }
+    public double SellingYield { get; set; }
+    public double BuyVolume { get; set; } = 50000000;
+    public double SellVolume { get; set; } = 50000000;
     [ForeignKey("Institution")]
     public string InstitutionId { get; set; } = null!;
     public Institution Institution { get; set; } = null!;
